Refactor ConvertSinglePage component and tests
- Removed the "Cancel Task" button from the modal and updated related tests.
- Added new ConversionForm, ConversionModal, and ConversionStatus components for better separation of concerns.
- Introduced a custom hook, useConversion, to manage conversion logic and state.
- Updated modal implementation to improve visibility control and styling.
- Added utility functions for status handling and navigation.
- Removed deprecated modal component and associated tests.
