+++ web/src/components/ConvertSinglePage/index.tsx
@@ -1,8 +1,8 @@
-import { Payload, UIField } from 'payload'
+import { UIField } from 'payload'
 import ConvertSinglePageClient from './client'
 
 async function ConvertSinglePage({ field }: { field: UIField }) {
-  return <ConvertSinglePageClient field={field} />;
+  return <ConvertSinglePageClient field={field} />
 }
 
 export default ConvertSinglePage
